feat: 정렬 기능 추가 (STAR: 평점 높은순, REVIEW: 리뷰 갯수 많은순, LATEST: product 등록순)

+++ src/main/java/com/team5/catdogeats/products/domain/dto/MyProductResponseDto.java
@@ -2,13 +2,11 @@ package com.team5.catdogeats.products.domain.dto;
 
 import com.team5.catdogeats.storage.domain.dto.ProductImageResponseDto;
 
-import java.util.List;
-
 // 특정 sellerId로 상품 조회시 response
 public record MyProductResponseDto(
     String productId,
     String productName,
-    int reviewCount,
+    long reviewCount,
     double averageStar,
     ProductImageResponseDto image
 ) {
